Test: MemberWithdrawServiceTest 클래스 추가
