+++ src/test/java/com/dnd/runus/application/oauth/OauthServiceTest.java
@@ -5,18 +5,8 @@ import com.dnd.runus.auth.oidc.provider.OidcProvider;
 import com.dnd.runus.auth.oidc.provider.OidcProviderRegistry;
 import com.dnd.runus.auth.token.TokenProviderModule;
 import com.dnd.runus.auth.token.dto.AuthTokenDto;
-import com.dnd.runus.domain.badge.BadgeAchievementRepository;
-import com.dnd.runus.domain.challenge.achievement.ChallengeAchievementPercentageRepository;
-import com.dnd.runus.domain.challenge.achievement.ChallengeAchievementRepository;
-import com.dnd.runus.domain.common.Coordinate;
-import com.dnd.runus.domain.common.Pace;
-import com.dnd.runus.domain.goalAchievement.GoalAchievementRepository;
 import com.dnd.runus.domain.member.*;
-import com.dnd.runus.domain.running.RunningRecord;
-import com.dnd.runus.domain.running.RunningRecordRepository;
-import com.dnd.runus.domain.scale.ScaleAchievementRepository;
 import com.dnd.runus.global.constant.MemberRole;
-import com.dnd.runus.global.constant.RunningEmoji;
 import com.dnd.runus.global.constant.SocialType;
 import com.dnd.runus.global.exception.BusinessException;
 import com.dnd.runus.global.exception.NotFoundException;
@@ -34,11 +24,9 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.context.ApplicationEventPublisher;
 
-import java.time.Duration;
 import java.time.OffsetDateTime;
-import java.util.Collections;
-import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.CountDownLatch;
 
@@ -70,22 +58,7 @@ class OauthServiceTest {
     private MemberLevelRepository memberLevelRepository;
 
     @Mock
-    private BadgeAchievementRepository badgeAchievementRepository;
-
-    @Mock
-    private RunningRecordRepository runningRecordRepository;
-
-    @Mock
-    private ChallengeAchievementRepository challengeAchievementRepository;
-
-    @Mock
-    private GoalAchievementRepository goalAchievementRepository;
-
-    @Mock
-    private ChallengeAchievementPercentageRepository challengeAchievementPercentageRepository;
-
-    @Mock
-    private ScaleAchievementRepository scaleAchievementRepository;
+    private ApplicationEventPublisher eventPublisher;
 
     private Member member;
 
@@ -303,116 +276,6 @@ class OauthServiceTest {
 
             // when, then
             assertThrows(AuthException.class, () -> oauthService.revokeOauth(member.memberId(), request));
-        }
-    }
-
-    @Nested
-    @ExtendWith(MockitoExtension.class)
-    @DisplayName("회원 탈퇴를 위한 멤버 데이터 삭제")
-    class DeleteAllMemberTest {
-
-        @DisplayName("회원 삭제: 회원이 존재하지 않으면 NotFoundException을 발생하한다.")
-        @Test
-        public void testDeleteAllDataAboutMember_MemberNotFound() {
-            given(memberRepository.findById(member.memberId())).willReturn(Optional.empty());
-
-            assertThrows(NotFoundException.class, () -> oauthService.deleteAllDataAboutMember(member.memberId()));
-        }
-
-        @DisplayName("회원 삭제 : running_record 존재 X")
-        @Test
-        void testDeleteAllDataAboutMember_NoRunningRecords() {
-            // given
-            given(memberRepository.findById(member.memberId())).willReturn(Optional.of(member));
-            given(runningRecordRepository.findByMember(member)).willReturn(Collections.emptyList());
-
-            // when
-            oauthService.deleteAllDataAboutMember(member.memberId());
-
-            // then
-            then(memberLevelRepository).should().deleteByMemberId(member.memberId());
-            then(badgeAchievementRepository).should().deleteByMemberId(member.memberId());
-            then(scaleAchievementRepository).should().deleteByMemberId(member.memberId());
-            then(socialProfileRepository).should().deleteByMemberId(member.memberId());
-            then(memberRepository).should().deleteById(member.memberId());
-        }
-
-        @DisplayName("회원 삭제 : running_record 존재, challenge_achievement 존재 X")
-        @Test
-        void testDeleteAllDataAboutMember_WithRunningRecords_NoChallengeAchievement() {
-            // given
-            List<RunningRecord> runningRecords = List.of(new RunningRecord(
-                    1L,
-                    member,
-                    1_100_000,
-                    Duration.ofHours(12).plusMinutes(23).plusSeconds(56),
-                    1,
-                    new Pace(5, 11),
-                    OffsetDateTime.now(),
-                    OffsetDateTime.now().plusHours(1),
-                    List.of(new Coordinate(1, 2, 3), new Coordinate(4, 5, 6)),
-                    "start location",
-                    "end location",
-                    RunningEmoji.SOSO));
-
-            given(memberRepository.findById(member.memberId())).willReturn(Optional.of(member));
-            given(runningRecordRepository.findByMember(member)).willReturn(runningRecords);
-            given(challengeAchievementRepository.findIdsByRunningRecords(runningRecords))
-                    .willReturn(Collections.emptyList());
-
-            // when
-            oauthService.deleteAllDataAboutMember(member.memberId());
-
-            // then
-            then(memberLevelRepository).should().deleteByMemberId(member.memberId());
-            then(badgeAchievementRepository).should().deleteByMemberId(member.memberId());
-            then(scaleAchievementRepository).should().deleteByMemberId(member.memberId());
-            then(socialProfileRepository).should().deleteByMemberId(member.memberId());
-            then(goalAchievementRepository).should().deleteByRunningRecords(runningRecords);
-            then(runningRecordRepository).should().deleteByMemberId(member.memberId());
-            then(memberRepository).should().deleteById(member.memberId());
-        }
-
-        @DisplayName("회원 삭제 : running_record, challenge_achievement 존재")
-        @Test
-        void testDeleteAllDataAboutMember_WithRunningRecords_WithChallengeAchievement() {
-            // given
-            List<RunningRecord> runningRecords = List.of(new RunningRecord(
-                    1L,
-                    member,
-                    1_100_000,
-                    Duration.ofHours(12).plusMinutes(23).plusSeconds(56),
-                    1,
-                    new Pace(5, 11),
-                    OffsetDateTime.now(),
-                    OffsetDateTime.now().plusHours(1),
-                    List.of(new Coordinate(1, 2, 3), new Coordinate(4, 5, 6)),
-                    "start location",
-                    "end location",
-                    RunningEmoji.SOSO));
-
-            List<Long> challengeAchievementIds = List.of(1L);
-
-            given(memberRepository.findById(member.memberId())).willReturn(Optional.of(member));
-            given(runningRecordRepository.findByMember(member)).willReturn(runningRecords);
-            given(challengeAchievementRepository.findIdsByRunningRecords(runningRecords))
-                    .willReturn(challengeAchievementIds);
-
-            // when
-            oauthService.deleteAllDataAboutMember(member.memberId());
-
-            // then
-            then(memberLevelRepository).should().deleteByMemberId(member.memberId());
-            then(badgeAchievementRepository).should().deleteByMemberId(member.memberId());
-            then(scaleAchievementRepository).should().deleteByMemberId(member.memberId());
-            then(socialProfileRepository).should().deleteByMemberId(member.memberId());
-            then(goalAchievementRepository).should().deleteByRunningRecords(runningRecords);
-            then(challengeAchievementPercentageRepository)
-                    .should()
-                    .deleteByChallengeAchievementIds(challengeAchievementIds);
-            then(challengeAchievementRepository).should().deleteByIds(challengeAchievementIds);
-            then(runningRecordRepository).should().deleteByMemberId(member.memberId());
-            then(memberRepository).should().deleteById(member.memberId());
         }
     }
 }
